Seems to work for higher rates

+++ rtp.h
@@ -25,9 +25,9 @@ int have_timestamp_timing_information(rtsp_conn_info *conn);
 
 int get_frame_play_time(int64_t timestamp, int sample_ratio, uint64_t *time_to_play);
 
-int frame_to_local_time(int64_t timestamp, uint64_t *time, rtsp_conn_info *conn);
-int local_time_to_frame(uint64_t time, int64_t *frame, rtsp_conn_info *conn);
+int frame_to_local_time(uint32_t timestamp, uint64_t *time, rtsp_conn_info *conn);
+int local_time_to_frame(uint64_t time, uint32_t *frame, rtsp_conn_info *conn);
 
-int sanitised_source_rate_information(int64_t *frames, uint64_t *time, rtsp_conn_info *conn);
+int sanitised_source_rate_information(uint32_t *frames, uint64_t *time, rtsp_conn_info *conn);
 
 #endif // _RTP_H
